+++ src/test/resources/io/github/zlika/reproducible/ObjectFactory-glassfish.java
@@ -1,0 +1,61 @@
+//
+// This file was generated by the Eclipse Implementation of JAXB, v2.3.6 
+// See https://eclipse-ee4j.github.io/jaxb-ri 
+// Any modifications to this file will be lost upon recompilation of the source schema. 
+// Generated on: 2022.07.20 at 02:38:48 PM CEST 
+//
+
+package test;
+
+import jakarta.xml.bind.annotation.XmlRegistry;
+
+/**
+ * This object contains factory methods for each 
+ * Java content interface and Java element interface 
+ * generated in the test package.
+ * <p>An ObjectFactory allows you to programatically 
+ * construct new instances of the Java representation 
+ * for XML content. The Java representation of XML 
+ * content can consist of schema derived interfaces 
+ * and classes representing the binding of schema 
+ * type definitions, element declarations and model 
+ * groups.  Factory methods for each of these are 
+ * provided in this class.
+ * 
+ */
+@XmlRegistry
+public class ObjectFactory {
+
+
+    /**
+     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: ch.post.it.evoting.tools.commons.model.gravity.evotingConfig
+     *
+     */
+    public ObjectFactory() {
+    }
+
+    /**
+     * Create an instance of {@link ListUnionDescriptionType }
+     *
+     */
+    public ListUnionDescriptionType createListUnionDescriptionType() {
+        return new ListUnionDescriptionType();
+    }
+
+    /**
+     * Create an instance of {@link PersonType }
+     * 
+     */
+    public PersonType createPersonType() {
+        return new PersonType();
+    }
+
+    /**
+     * Create an instance of {@link BallotQuestionType }
+     * 
+     */
+    public BallotQuestionType createBallotQuestionType() {
+        return new BallotQuestionType();
+    }
+
+}
